refactor(ast): streamline AST module structure
Simplify AST codebase by removing redundancy and better organizing code:

## Changes Made

### Documentation Improvements
- Move architecture docs from range.rs to ast.rs module level
- Keeps range.rs focused on type-specific documentation
- Better organization and discoverability

### Code Deduplication
- Remove duplicate node_type() and display_label() methods from ContentItem
- Add AstNode trait imports where needed
- Code now properly uses trait methods consistently

### Module Consolidation
- Merge snapshot_visitor.rs into snapshot.rs
- Tests colocated with implementation
- Remove lookup.rs, inline functionality into Document methods
- Add backwards-compatible wrapper functions for existing API

### Cleanup
- Remove deprecated byte_range_to_location() function
- Update imports across formatters (treeviz, tag) and lex-viewer

## Results
- Saves ~230-260 lines (11-13% reduction)
- All 251 library tests passing
- No breaking changes
- Better code organization
